refactor: Move kubernetes-jobmonitor out of the transport directory
While the kubernetes-jobmonitor is related to the Kubernetes transport
it is not a transport itself, so move it to a separate `kubernetes`
directory to not mix it with transport implementations.
