+++ settings.gradle.kts
@@ -34,6 +34,7 @@ include(":config:secret-file")
 include(":config:spi")
 include(":core")
 include(":dao")
+include(":kubernetes:jobmonitor")
 include(":logaccess:spi")
 include(":logaccess:loki")
 include(":model")
@@ -52,7 +53,6 @@ include(":storage:s3")
 include(":storage:spi")
 include(":transport:activemqartemis")
 include(":transport:kubernetes")
-include(":transport:kubernetes-jobmonitor")
 include(":transport:rabbitmq")
 include(":transport:spi")
 include(":transport:sqs")
